Organize EC2 classes in Amazon provider
