+++ airflow/providers/amazon/aws/sensors/ec2_instance_state.py
@@ -15,51 +15,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-#
+"""This module is deprecated. Please use :mod:`airflow.providers.amazon.aws.sensors.ec2`."""
 
-from typing import Optional
+import warnings
 
-from airflow.providers.amazon.aws.hooks.ec2 import EC2Hook
-from airflow.sensors.base import BaseSensorOperator
+from airflow.providers.amazon.aws.sensors.ec2 import EC2InstanceStateSensor # noqa
 
-
-class EC2InstanceStateSensor(BaseSensorOperator):
-    """
-    Check the state of the AWS EC2 instance until
-    state of the instance become equal to the target state.
-
-    :param target_state: target state of instance
-    :type target_state: str
-    :param instance_id: id of the AWS EC2 instance
-    :type instance_id: str
-    :param region_name: (optional) aws region name associated with the client
-    :type region_name: Optional[str]
-    """
-
-    template_fields = ("target_state", "instance_id", "region_name")
-    ui_color = "#cc8811"
-    ui_fgcolor = "#ffffff"
-    valid_states = ["running", "stopped", "terminated"]
-
-    def __init__(
-        self,
-        *,
-        target_state: str,
-        instance_id: str,
-        aws_conn_id: str = "aws_default",
-        region_name: Optional[str] = None,
-        **kwargs,
-    ):
-        if target_state not in self.valid_states:
-            raise ValueError(f"Invalid target_state: {target_state}")
-        super().__init__(**kwargs)
-        self.target_state = target_state
-        self.instance_id = instance_id
-        self.aws_conn_id = aws_conn_id
-        self.region_name = region_name
-
-    def poke(self, context):
-        ec2_hook = EC2Hook(aws_conn_id=self.aws_conn_id, region_name=self.region_name)
-        instance_state = ec2_hook.get_instance_state(instance_id=self.instance_id)
-        self.log.info("instance state: %s", instance_state)
-        return instance_state == self.target_state
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.ec2`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
